Update app.py
comment out a piece of code that is causing an issue in the app the run. 

This is a temp fix to get the app working and to be able to go into the next lesson of the Udacity course, but this needs to be fixed in a later stage in a proper way.